перенёс функцию создания фильтра в модуль
