+++ src/main.js
@@ -1,36 +1,6 @@
-'use strict';
-const filtersNames = [
-  `all`,
-  `overdue`,
-  `today`,
-  `favorites`,
-  `repeating`,
-  `tags`,
-  `archive`
-];
+import {generateFilter, filtersNames} from './generate-filter';
 
-const MAX_CARDS_NUMBERS = 15;
-const MIN_CARDS_NUMBERS = 1;
 const START_CARDS_COUNT = 7;
-
-const getRandomNumber = (min, max) => {
-  let rand = min + Math.random() * (max + 1 - min);
-  rand = Math.floor(rand);
-  return rand;
-};
-
-const generateFilter = (count) => {
-  const cardNumbers = getRandomNumber(MIN_CARDS_NUMBERS, MAX_CARDS_NUMBERS);
-  return `<input
-            type="radio"
-            id="filter__${filtersNames[count]}"
-            class="filter__input visually-hidden"
-            name="filter"/>
-    <label for="filter__${filtersNames[count]}" class="filter__label">
-      ${filtersNames[count]}
-      <span class="filter__all-count">${cardNumbers}</span>
-    </label>`;
-};
 
 const generateCard = () => {
   return `<article class="card card--pink card--repeat">
