+++ src/utilities.js
@@ -1,0 +1,5 @@
+export const getRandomNumber = (min, max) => {
+  let rand = min + Math.random() * (max + 1 - min);
+  rand = Math.floor(rand);
+  return rand;
+};
